two direct endpoints for the protvista component to fetch data from Pharos
